requirements.txt
psycopg2==2.9.10
pytz==2025.1
RPi.GPIO==0.7.1
